gen: carry the panic message with the status code and drop LockOSThread
A caught Zig panic publishes its message in a sequence-tagged slot table
and returns -(256 + sequence); {prefix}_caught_panic_message(code) reads
it from any thread, so generated functions no longer pin the OS thread.
Projection and snapshot accessors return int32_t and use the same code
range, which removes the last thread-local message read. The purego
public file keeps its runtime import only for the cancel-flag pinner.

+++ examples/07-event-queue/go-purego/internal/lifecycle/lifecycle_gen.go
@@ -115,7 +115,7 @@ func (err *CallbackError) Unwrap() error        { return err.Err }
 
 type StatusError struct {
 	Operation string
-	Status    uint8
+	Status    int32
 }
 
 func (err *StatusError) Error() string {
